+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ### Added
 - is_remote:  Determines if the remote git repository exists.
 
+### Changed
+- merge:  Replaced is_remote_branch() call with is_remote() call.
+
 
 ## [0.0.1] - 2019-03-19
 - Initial creation.